server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
httpx==0.28.1
rembg==2.0.59
pillow==11.1.0
numpy==2.2.2
duckduckgo_search==7.5.1
